requirements.txt
colorthief==0.2.1
joblib==1.5.1
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
scikit-learn==1.6.1
scipy==1.15.3
threadpoolctl==3.6.0
